+++ lib/flutter_native_video_trimmer.dart
@@ -1,5 +1,6 @@
 library flutter_native_video_trimmer;
 
+export 'src/models/models.dart';
 export 'src/video_trimmer.dart';
 export 'src/video_trimmer_platform_interface.dart';
 export 'src/video_trimmer_method_channel.dart';